+++ pkg/server/config/config.go
@@ -99,7 +99,6 @@ const (
 	DefaultSharedIteratorMaxAdmissionTime = 10 * time.Second
 	DefaultSharedIteratorMaxIdleTime      = 1 * time.Second
 
-	DefaultPlannerInitialGuess      = 10 * time.Millisecond
 	DefaultPlannerEvictionThreshold = 0
 	DefaultPlannerCleanupInterval   = 0
 
@@ -298,7 +297,6 @@ type AccessControlConfig struct {
 }
 
 type PlannerConfig struct {
-	InitialGuess      time.Duration
 	EvictionThreshold time.Duration
 	CleanupInterval   time.Duration
 }
@@ -822,7 +820,6 @@ func DefaultConfig() *Config {
 		RequestTimeout:                DefaultRequestTimeout,
 		ContextPropagationToDatastore: false,
 		Planner: PlannerConfig{
-			InitialGuess:      DefaultPlannerInitialGuess,
 			EvictionThreshold: DefaultPlannerEvictionThreshold,
 			CleanupInterval:   DefaultPlannerCleanupInterval,
 		},