+++ pkg/server/server.go
@@ -238,6 +238,8 @@ type Server struct {
 	singleflightGroup *singleflight.Group
 
 	planner *planner.Planner
+
+	requestTimeout time.Duration
 }
 
 type OpenFGAServiceV1Option func(s *Server)
@@ -582,6 +584,12 @@ func WithPlanner(planner *planner.Planner) OpenFGAServiceV1Option {
 	}
 }
 
+func WithRequestTimeout(timeout time.Duration) OpenFGAServiceV1Option {
+	return func(s *Server) {
+		s.requestTimeout = timeout
+	}
+}
+
 // MustNewServerWithOpts see NewServerWithOpts.
 func MustNewServerWithOpts(opts ...OpenFGAServiceV1Option) *Server {
 	s, err := NewServerWithOpts(opts...)
@@ -873,10 +881,10 @@ func NewServerWithOpts(opts ...OpenFGAServiceV1Option) (*Server, error) {
 		singleflightGroup: &singleflight.Group{},
 		authorizer:        authz.NewAuthorizerNoop(),
 		planner: planner.New(&planner.Config{
-			InitialGuess:      serverconfig.DefaultPlannerInitialGuess,
 			EvictionThreshold: serverconfig.DefaultPlannerEvictionThreshold,
 			CleanupInterval:   serverconfig.DefaultPlannerCleanupInterval,
 		}),
+		requestTimeout: serverconfig.DefaultRequestTimeout,
 	}
 
 	for _, opt := range opts {
@@ -973,6 +981,8 @@ func NewServerWithOpts(opts ...OpenFGAServiceV1Option) (*Server, error) {
 			graph.WithResolveNodeBreadthLimit(s.resolveNodeBreadthLimit),
 			graph.WithMaxResolutionDepth(s.resolveNodeLimit),
 			graph.WithPlanner(s.planner),
+			graph.WithUpstreamTimeout(s.requestTimeout),
+			graph.WithLocalCheckerLogger(s.logger),
 		}...),
 		graph.WithLocalShadowCheckerOpts([]graph.LocalCheckerOption{
 			graph.WithResolveNodeBreadthLimit(s.resolveNodeBreadthLimit),