+++ CHANGELOG.md
@@ -12,6 +12,7 @@ Try to keep listed changes to a concise bulleted list of simple explanations of 
 
 ### Changed
 - Bumped the version of `openfga/language/pkg` to a version of the weighted graph that includes recursive relation detection. [#2716](https://github.com/openfga/openfga/pull/2716)
+- Log the reason on server failure start [#2703](https://github.com/openfga/openfga/pull/2703)
 
 ## [1.10.1] - 2025-09-22
 ### Fixed